BUG: DataFrame._item_cache not cleared on on .copy() (#33299)
